+++ src/main/java/com/example/book/repository/HospitalRepository.java
@@ -3,5 +3,5 @@ package com.example.book.repository;
 import com.example.book.domain.entity.Hospital;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface HospitalRepository extends JpaRepository<Hospital, Integer> {
+public interface HospitalRepository extends JpaRepository<Hospital, Long> {
 }
